+++ Expectativas_de_Mercado.Model/ValueObjects/DesvioPadrao.cs
@@ -1,9 +1,9 @@
 ﻿namespace Expectativas_de_Mercado.Model.ValueObjects;
 public record DesvioPadrao
 {
-    public double Value { get; set; }
-    public static implicit operator double(DesvioPadrao d) => d.Value;
-    public static implicit operator DesvioPadrao(double  value) => new DesvioPadrao(value);
+    public decimal Value { get; set; }
+    public static implicit operator decimal(DesvioPadrao d) => d.Value;
+    public static implicit operator DesvioPadrao(decimal  value) => new DesvioPadrao(value);
     public DesvioPadrao() { }
-    public DesvioPadrao(double value) => this.Value = value;
+    public DesvioPadrao(decimal value) => this.Value = value;
 }
