+++ Expectativas_de_Mercado.Model/ValueObjects/Media.cs
@@ -1,9 +1,9 @@
 ﻿namespace Expectativas_de_Mercado.Model.ValueObjects;
 public record Media 
 {
-    public double Value { get; set; }
-    public static implicit operator double(Media d) => d.Value;
-    public static implicit operator Media(double  value) => new Media(value);
+    public decimal Value { get; set; }
+    public static implicit operator decimal(Media d) => d.Value;
+    public static implicit operator Media(decimal  value) => new Media(value);
     public Media() { }
-    public Media(double value) => this.Value = value;
+    public Media(decimal value) => this.Value = value;
 }
