+++ Expectativas_de_Mercado.Model/ValueObjects/Mediana.cs
@@ -1,9 +1,9 @@
 ﻿namespace Expectativas_de_Mercado.Model.ValueObjects;
 public record Mediana 
 {
-    public double Value { get; set; }
-    public static implicit operator double (Mediana d) => d.Value;
-    public static implicit operator Mediana(double  value) => new Mediana(value);
+    public decimal Value { get; set; }
+    public static implicit operator decimal (Mediana d) => d.Value;
+    public static implicit operator Mediana(decimal  value) => new Mediana(value);
     public Mediana() { }
-    public Mediana(double  value) => this.Value = value;
+    public Mediana(decimal  value) => this.Value = value;
 }
